Auto reconnect fixes and SDK fixes for Firmware 73 Pencils
- Move recent work from develop (244025a)
- Merge FiftyThreeSdk code from develop (f03d4d6cedbafa8f62335a75fa60670acbb58692)
- Fix build breaks LineTest,OldSDK,CanvasApp

+++ FiftyThreeSdk/FiftyThreeSdk/CBCharacteristic+Helpers.m
@@ -2,7 +2,7 @@
 //  CBCharacteristic+Helpers.m
 //  FiftyThreeSdk
 //
-//  Copyright (c) 2013 FiftyThree, Inc. All rights reserved.
+//  Copyright (c) 2014 FiftyThree, Inc. All rights reserved.
 //
 
 #import "CBCharacteristic+Helpers.h"
@@ -12,7 +12,7 @@
 - (NSData *)removeTrailingZeros:(NSData *)data
 {
     NSData *value = self.value;
-    int length = value.length;
+    NSUInteger length = value.length;
     for (; length > 0; length--)
     {
         if (((uint8_t *)value.bytes)[length - 1] != '\0')
